refactor(pkg): renamed package to michaelyi
