+++ backend/config/checkstyle/checkstyle.xml
@@ -83,7 +83,9 @@
         <!-- Checks for Naming Conventions.                  -->
         <!-- See http://checkstyle.sf.net/config_naming.html -->
         <module name="ConstantName"/>
-        <module name="LocalFinalVariableName"/>
+        <module name="LocalFinalVariableName">
+            <property name="format" value="^[A-Z][A-Z0-9]*$"/>  
+        </module>
         <module name="LocalVariableName"/>
         <module name="MemberName"/>
         <module name="MethodName"/>
