--- backend/src/main/java/com/michaelhyi/auth/AuthLoginRequest.java
@@ -1,6 +1,0 @@
-package com.michaelhyi.auth;
-
-public record AuthLoginRequest(
-        String email
-) {
-}
