--- backend/src/main/java/com/michaelhyi/auth/UnauthorizedException.java
@@ -1,7 +1,0 @@
-package com.michaelhyi.auth;
-
-public class UnauthorizedException extends RuntimeException {
-    public UnauthorizedException() {
-        super("Unauthorized.");
-    }
-}
