--- backend/src/main/java/com/michaelhyi/auth/UserNotFoundException.java
@@ -1,7 +1,0 @@
-package com.michaelhyi.auth;
-
-public class UserNotFoundException extends RuntimeException {
-    public UserNotFoundException() {
-        super("User not found.");
-    }
-}
