--- backend/src/main/java/com/michaelhyi/post/PostRepository.java
@@ -1,9 +1,0 @@
-package com.michaelhyi.post;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import jakarta.transaction.Transactional;
-
-@Transactional
-public interface PostRepository extends JpaRepository<Post, String> {
-}
