--- backend/src/main/java/com/michaelhyi/s3/S3Config.java
@@ -1,33 +1,0 @@
-package com.michaelhyi.s3;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
-import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider;
-import software.amazon.awssdk.regions.Region;
-import software.amazon.awssdk.services.s3.S3Client;
-
-@Configuration
-public class S3Config {
-    @Value("${aws.access-key}")
-    private String accessKey;
-
-    @Value("${aws.secret-key}")
-    private String secretKey;
-
-    @Bean
-    public S3Client s3Client() {
-        AwsBasicCredentials credentials = AwsBasicCredentials
-                .create(accessKey, secretKey);
-
-        return S3Client
-                .builder()
-                .credentialsProvider(
-                        StaticCredentialsProvider
-                                .create(credentials))
-                .region(Region.of("us-west-2"))
-                .build();
-    }
-}
