--- backend/src/main/java/com/michaelhyi/security/AppConfig.java
@@ -1,43 +1,0 @@
-package com.michaelhyi.security;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.AuthenticationProvider;
-import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
-import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-
-import com.michaelhyi.user.UserRepository;
-
-import lombok.RequiredArgsConstructor;
-
-@Configuration
-@RequiredArgsConstructor
-public class AppConfig {
-    private final UserRepository repository;
-
-    @Bean
-    public AuthenticationManager authenticationManager(
-            AuthenticationConfiguration config
-    ) throws Exception {
-        return config.getAuthenticationManager();
-    }
-
-    @Bean
-    public AuthenticationProvider authenticationProvider() {
-        DaoAuthenticationProvider provider =
-                new DaoAuthenticationProvider();
-        provider.setUserDetailsService(userDetailsService());
-        return provider;
-    }
-
-    @Bean
-    public UserDetailsService userDetailsService() {
-        return username -> repository.findById(username)
-                .orElseThrow(() ->
-                        new UsernameNotFoundException("User not found.")
-                );
-    }
-}
