--- backend/src/main/java/com/michaelhyi/user/UserRepository.java
@@ -1,9 +1,0 @@
-package com.michaelhyi.user;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import jakarta.transaction.Transactional;
-
-@Transactional
-public interface UserRepository extends JpaRepository<User, String> {
-}
