+++ backend/src/main/java/com/michaelyi/auth/AuthExceptionHandler.java
@@ -1,0 +1,23 @@
+package com.michaelyi.auth;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+@ControllerAdvice
+public class AuthExceptionHandler {
+    @ExceptionHandler(UnauthorizedException.class)
+    public ResponseEntity<String> handleException(
+            UnauthorizedException e
+    ) {
+        return new ResponseEntity<>(e.getMessage(), HttpStatus.UNAUTHORIZED);
+    }
+
+    @ExceptionHandler(UserNotFoundException.class)
+    public ResponseEntity<String> handleException(
+            UserNotFoundException e
+    ) {
+        return new ResponseEntity<>(e.getMessage(), HttpStatus.NOT_FOUND);
+    }
+}
