+++ backend/src/main/java/com/michaelyi/auth/AuthLoginRequest.java
@@ -1,0 +1,6 @@
+package com.michaelyi.auth;
+
+public record AuthLoginRequest(
+        String email
+) {
+}
