+++ backend/src/main/java/com/michaelyi/auth/UnauthorizedException.java
@@ -1,0 +1,7 @@
+package com.michaelyi.auth;
+
+public class UnauthorizedException extends RuntimeException {
+    public UnauthorizedException() {
+        super("Unauthorized.");
+    }
+}
