+++ backend/src/main/java/com/michaelyi/auth/UserNotFoundException.java
@@ -1,0 +1,7 @@
+package com.michaelyi.auth;
+
+public class UserNotFoundException extends RuntimeException {
+    public UserNotFoundException() {
+        super("User not found.");
+    }
+}
