+++ backend/src/main/java/com/michaelyi/post/PostRepository.java
@@ -1,0 +1,9 @@
+package com.michaelyi.post;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import jakarta.transaction.Transactional;
+
+@Transactional
+public interface PostRepository extends JpaRepository<Post, String> {
+}
